Enhance cli args handling: add getopt_long
Switched to using getopt_long instead of getopt.

Remove function implementation from parse-util header file, and added
parse_port function

+++ src/orch/controller.h
@@ -3,6 +3,6 @@
 
 #include "../common/dbus.h"
 
-int controller_setup(int accept_fd, int port, sd_event *event, sd_event_source *event_source);
+int controller_setup(int accept_fd, uint16_t port, sd_event *event, sd_event_source *event_source);
 
 #endif
